Rename tools directory to tool
