+++ src/app/page.tsx
@@ -1,5 +1,5 @@
 import TechnologySection from '@/components/technology/TechnologySection';
-import ToolSection from '@/components/tools/ToolSection';
+import ToolSection from '@/components/tool/ToolSection';
 import SocialSection from '@/components/social/SocialSection';
 import ProjectSection from '@/components/projects/ProjectSection';
 import HomeHeroSection from '@/components/HomeHeroSection';
