+++ src/components/tool/ToolHeroSection.tsx
